+++ packages/api/lib/utils.js
@@ -78,36 +78,10 @@ function getGranuleProductVolume(granuleFiles = []) {
     .reduce((x, y) => x + y, 0);
 }
 
-/**
- * Find a property name in an object in a case-insensitive manner
- *
- * @param {Object} obj - the object to be searched
- * @param {string} keyArg - the name of the key to find
- * @returns {string|undefined} - the name of the matching key, or undefined if
- *   none was found
- */
-function findCaseInsensitiveKey(obj, keyArg) {
-  const keys = Object.keys(obj);
-  return keys.find((key) => key.toLowerCase() === keyArg.toLowerCase());
-}
-
-/**
- * Find a property value in an object in a case-insensitive manner
- *
- * @param {Object} obj - the object to be searched
- * @param {string} keyArg - the name of the key to find
- * @returns {*} the matching value
- */
-function findCaseInsensitiveValue(obj, keyArg) {
-  return obj[findCaseInsensitiveKey(obj, keyArg)];
-}
-
 module.exports = {
   deconstructCollectionId,
   errorify,
   extractDate,
-  findCaseInsensitiveKey,
-  findCaseInsensitiveValue,
   getGranuleProductVolume,
   parseException
 };